Extract useGetSchemas
